Title: 가장 많이 겹치는 구간 2, Time: 17ms, Memory: 0MB, Status: Runtime Error - Codetree

+++ 260314/가장 많이 겹치는 구간 2/section-with-maximum-overlap-2.cpp
@@ -19,7 +19,7 @@ int main() {
         mx = max(mx,x2[i]);
     }
 
-    for(int i = 1 ; i <= mx ; i++)
+    for(long long i = 1 ; i <= mx ; i++)
     {
         sum+=arr[i];
         ans = max(ans,sum);
